create separate newtype for producer trait
